third_party/WebKit/LayoutTests/fast/dom/HTMLDialogElement/inert-does-not-match-disabled-selector-expected.txt

Tests that inert elements do not match the :disabled selector.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS color is 'rgb(0, 128, 0)'
PASS successfullyParsed is true

TEST COMPLETE
The test passes if this is in green.
